ch(feedback): implement back-end feed-back
 -separate routes
 -decrease if statements by implementing try and catch brokes
 -delete dummy data from modals
 -separate controllers
 -separate tests
 -add helpers
 -use await for checking if a user exists

[Finishes #169434240]
