+++ Server/Model/entriesModel.js
@@ -1,24 +1,2 @@
-const entries = [
-    {
-        id: 1,
-        createdBy: 'systemadmin@gmail.com',
-        createdOn: 'October 15, 2019 9:20 PM',
-        title: 'My First Program',
-        description: 'Hello world with c programming',
-    },
-    {
-        id: 2,
-        createdBy: 'johnsmith@gmail.com',
-        createdOn: 'October 15, 2019 9:20 PM',
-        title: 'Le 21 dec 2010',
-        description: 'Hello world',
-    },
-    {
-        id: 3,
-        createdBy: 'andrewpeter@gmail.com',
-        createdOn: 'October 15, 2019 9:20 PM',
-        title: 'My birthday',
-        description: 'Hello world',
-    },
-];
+const entries = [];
 export default entries;
