+++ Server/Model/testEntry.js
@@ -1,0 +1,25 @@
+const testEntry = [
+    {
+        title: 'My title',
+        description: 'My description',
+    }, {
+        title: true,
+        description: 'hey there',
+    }, {
+        title: 'Le 21 dec 2019',
+        description: 'my new diary',
+    }, {
+        entryId: 1,
+    }, {
+        entryId: -1,
+    }, {
+        entryId: 500,
+    }, {
+        token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJmaXJzdE5hbWUiOiJ0ZXN0VXNlciIsImxhc3ROYW1lIjoidGVzdFVzZXIiLCJlbWFpbCI6InRlc3RVc2VyQGdtYWlsLmNvbSIsImlhdCI6OTk5OTk2NTM0MCwiZXhwIjo5OTk5OTcwMTQwfQ.5jE8K4QRSLk6cjwHsa7xOIDerRHB10v_30LnDg2DoKk',
+    }, {
+        token: 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJmaXJzdE5hbWUiOiJ0ZXN0VXNlciIsImxhc3ROYW1lIjoidGVzdFVzZXIiLCJlbWFpbCI6ImZhbHNlVGVzdFVzZXJAZ21haWwuY29tIiwiaWF0Ijo5OTk5OTY1MzQwLCJleHAiOjk5OTk5NzAxNDB9.9PVVkYa08yy5B5ZqPTFOksruU0gzWsE4LZ3zUVusHeQ',
+    }, {
+        token: '',
+    },
+];
+export default testEntry;
