+++ Server/Model/testUSer.js
@@ -1,0 +1,25 @@
+const testUsers = [{
+    firstName: 'testUser',
+    lastName: 'testUser',
+    email: 'testUser@gmail.com',
+    password: 'password',
+}, {
+    lastName: 'lastName',
+    firstName: '',
+    email: 'userEmail@gmail.com',
+    password: 'password',
+}, {
+    email: 'testUser@gmail.com',
+    password: 'password',
+}, {
+    email: 'testUser1@gmail.com',
+    password: 'password',
+}, {
+    email: 'testUser@gmail.com',
+    password: 'password1',
+}, {
+    email: 'testUser@gmail.com',
+    // password: 'password',
+},
+];
+export default testUsers;
