+++ Server/Model/userModel.js
@@ -1,22 +1,2 @@
-const users = [
-
-    {
-        firstName: 'system',
-        lastName: 'admin',
-        email: 'systemadmin@gmail.com',
-        password: '$2b$10$af95diBy/crnvEWxiA2r3u64S6osMVM0kg7mNIjm9AGwFkZYxa1Ni',
-    },
-    {
-        firstName: 'John',
-        lastName: 'Smith',
-        email: 'johnsmith@gmail.com',
-        password: '$2b$10$YJhS2kP82GXKAPwp9EULle4TLs0T05qPVOzKrCZACziadCmhXmQcq',
-    },
-    {
-        firstName: 'Andrew',
-        lastName: 'Peter',
-        email: 'andrewpeter@gmail.com',
-        password: '$2b$10$bV37mu7gaOE9Usmw4bg4DuRxFW1lKJKIKaVsLKjFnzBZVwUQFdshS',
-    },
-];
+const users = [];
 export default users;
