+++ Server/middleware/checkNewEntry.js
@@ -9,8 +9,7 @@ export const checkNewEntry = (req, res, next) => {
     if (schemasValidation.error) {
         const validationErrors = [];
         for (let i = 0; i < schemasValidation.error.details.length; i += 1) {
-            // eslint-disable-next-line quotes
-            validationErrors.push(schemasValidation.error.details[i].message.split('"').join(" "));
+            validationErrors.push(schemasValidation.error.details[i].message.split('"').join(' '));
         }
         return res.status(400).json({
             status: 400,
