+++ Server/middleware/checkParams.js
@@ -2,14 +2,13 @@ import Joi from 'joi';
 
 export const validateEntryParams = (req, res, next) => {
     const entryValidationParams = {
-        entryId: Joi.number().required(),
+        entryId: Joi.number().positive().required(),
     };
     const schemaValidation = Joi.validate(req.params, entryValidationParams);
     if (schemaValidation.error) {
         const validationErrors = [];
         for (let i = 0; i < schemaValidation.error.details.length; i += 1) {
-            // eslint-disable-next-line quotes
-            validationErrors.push(schemaValidation.error.details[i].message.split('"').join(" "));
+            validationErrors.push(schemaValidation.error.details[i].message.split('"').join(' '));
         }
         return res.status(400).json({
             status: 400,
