+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start": "babel-node index.js",
     "lint": "eslint ./",
-    "test": "nyc --reporter=html --reporter=text mocha --require @babel/register Server/Test/Test.js --exit",
+    "test": "nyc --reporter=html --reporter=text mocha --require @babel/register Server/Test/* --exit",
     "dev": "nodemon --exec babel-node index.js",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
